refactor: split stream tests
In order to improve the structuring of the tests and to separate tests
with jest mocks from the ones not using jest mocks.
